requirements.txt
asgiref==3.5.2
boto3==1.20.9
botocore==1.23.54
Django==3.2
django-storages==1.12.3
jmespath==0.10.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
